refactor(nznuxsir): update format code and folder name
